chore(language-js): clean up tests and coverage
Rename the integration test file of the `JavaScriptLanguagePlugin` to
more accurately reflect what is being tested.

Update the NYC configuration to ignore the barrel file of the tests
common directory because it skews the coverage negatively (specifically
function coverage) because the functions being exported are not tested
as being from the barrel file. Related to that, simplify an import path
for the unit tests for common expressions.