test(language-html): comments with newlines
Update the test suites for language-html expressions and embeds to
include tests for comments with newline characters.

For embeds, the list of test cases for the "comments" scenario of each
embed type is updated to include an example of a comment with a newline.

For expressions, the `buildHtmlComment` is upgraded to
`buildHtmlComments` to produce a list of comments instead of just one.
This new functionality is unit tested (in `builder.test.ts`) and used in
the unit test suite of each expression factory. On top of that the value
preset of comments (in `values.ts`) is updated to include a comment with
a newline, this affects all expression factory test suites as well as
the test suite for the common expressions.

This comment also contains two stylistict changes to the type definition
of the `valuePresets` object in accorance with the ESLint configuration.