test(language-html): non-standard comment closing for expressions
Update the test suites for the expression factories of the langauge-html
package to test for non-standard comment closings (`--!>` instead of
`-->`). This is achieved by adding it as a return value of the comment
builder test helper. The test suite for this helper is updated
accordingly. Additionally, the suites for the multi and single value
attribute expression factories have been updated to have fewer generated
test cases for the attribute-like comment scenario to avoid timeouts.