perf(language-utils): use Set to store languages in MultiLanguagePlugin
Update the abstract `MultiLanguagePlugin` to store the languages
supported by the language plugin in a set instead of an array to prevent
duplicates and the performance implications of duplicates downstream.

The version information of the class is updated accordingly.

Additionally, update two tests that assume the return value of
`getLanguages` is an array.