test(language-js): strings with extra quotes for query selector exprs
Add two test scenarios to the test suite for query selector expressions.
The first checks that a string containing a query selector that should
be detected works as expected if that string contains a quote character.
The second checks that an unrelated string with a quote character is
properly ignored and another string that does contain a query selector
is detected.