test(language-html): finding embeds with non-standard closing tags
Add test cases for finding both CSS and JavaScript embeds in HTML if the
closing <style> or <script> (resp.) tag is non-standard. In particular,
when the closing tag has "attributes".
This testcase is necessary because browsers will generally accept such
stylesheets/scripts (resp.) as valid and use them when displaying the
page. As such these stylesheets/scripts should be mangled.