test(cli): separate unit and integration stats
Refactor the stats code & tests of the CLI package such that all tests
are effectively unit tests - i.e. they only assume that dependencies
work, not how they work - as the refactoring needed to use dependency
injection or otherwise decouple the code from its dependencies is beyond
the scope of this branch.