refactor(core)!: let main function return an object
Refactor the main function to return an object (containing the mangled
files) rather than the mangled files by themselves. This is a preventive
measure meant to avoid a breaking change in the future when the main
function should return more than one value (e.g. a cache).

BREAKING CHANGE: The API of the core has changed, update how you use it.
For example:

    import webmangler from "webmangler";
    const originalFiles = [/* some files */];
    const options = {/* some options */};
    const files = webmangler(originalFiles, options);

becomes:

    import webmangler from "webmangler";
    const originalFiles = [/* some files */];
    const options = {/* some options */};
    const { files } = webmangler(originalFiles, options);