refactor(core): remove characters module
Remove the `characters.ts` module from the core. Since the `CharSet`s
were added to the mangler-utils (in v0.1.24) the various `CharSet`s are
no longer needed externally. Internally their use can be reduced to a
single location(*) - the `NameGenerator` class - as a default for embed
identifiers (embeds.ts:49) and intermediate mangle values
(replace.ts:16). Since both of these uses are for internal state (ids)
only, the character set can be more or less arbitrary. (to avoid issues
in the future, the optional `charSet` may be made required in the future
for clarity but this depends on the note(*) below). The `CharSet`
previously used by the embeds code was chosen with performance in mind
(it has a higher probability of finding a short unique string), the test
suite is updated accordingly.

The duplication of the character sets in the test code is justified by
the fact that 1) the reduces the amount of changes, and 2) it makes the
tests as they are self contained.

A note on seemingly unrelated changes:
- All type imports that referenced `characters.ts` now reference the
  types package.
- The `TestCase` type in `name-generator.test.ts` was updated due to
  formatting rules.
- The `name-generator.test.ts` now uses the non-deprecated
  `TestScenarios` type from the testing package.

(*): There's currently a default `CharSet` in the `MangleEgnineOptions`
parser function, this default is planned for removal in the future. The
reason for this is that manglers should be required to be explicit about
the `CharSet` they want to use.