test(language-utils): no `groupName` in `SingleGroupMangleExpression`
Update the `SingleGroupMangleExpression` in preparation for deprecating
the `groupName` in favor of a provided capturing group string. This is
not yet fully implemented, but the expected behaviour is documented and
tested.

A test case is added to verify that the `SingleGroupMangleExpression`
constructor throws when provided with a `patternTemplate` without the
provided capturing group string (and no `groupName`, at least for the
transition period). The scenarios are updated in accordance with the new
design. For each scenario, two test cases are ran:

1. With the new constructor, omitting the `groupName` and using the
   provided capturing group string.
2. With the old constructor where the provided capturing group string is
   replaced by the string "(?<g>%s)", which matching the old scenarios.

The previously called "missing group" test scenario is replaced by a
scenario more accurately reflecting what it tests (also, "missing group"
is a nonsensical test scenario without a provided `groupName`). What it
actually tests is that when a group that is not the find and replace
capturing is triggered, the `SingleGroupMangleExpression` ignores that
match.

As the `groupName` is no longer relevant to the tests, the `group` field
of the `TestCase` types has been removed.

**NOTE**:
- The `.replace(/\s/g, "")` is present in most scenarios because the
  scenarios cannot assume that whitespace is removed from the template
  by the `SingleGroupMangleExpression`.