refactor(core): make `charSet` for `NameGenerator` required
Update the `NameGenerator` to make the `charSet` input required instead
of optional, removing the fallback and simplifying the test suite. The
fallback was relied upon in two places, both of which now have their own
dedicated `CharSet` that's passed to the `NameGenerator`.

As the `NameGeneratorOptions` is touched anyway (to make `charSet`
required) , it was cleaned up a bit as well. First, it's been changed
from a `type` to an `interface` for consistency throughout the project.
Second, both options are now `readonly`, preventing the `NameGenerator`
from re-assigning them. Third the invalid link in the `reservedNames`
doc is fixed.

Related to the previous, for consistency with the rest of the project
the export statement for the `NameGenerator` is now separate from the
class definition.