feat(scripts): check flags in scripts
Add a utility script to work with flags that currently exposes a
function to check if the flags provided to a CLI are "valid" for that
CLI.

Each CLI script now uses this function to check the provided args
against their `FLAGS` object.

Additionally, the CLIs now consistently pass a sliced `process.argv`
value to the `main` function to prevent any usage of the `argv` before
it's sliced.