config: common definitions for test subdirectories
Add definitions for the test subdirectories to the values module and use
these values in configurations. In the Mocha config (`.mocharc.cjs`)
this replaces the hardcoded `testTypeDir` values. In the NYC config
(`nyc.config.cjs`) this is used in the exclusion pattern for the barrel
file of `__tests__/common` (*). In `get-filters.js` this is used in the
regular expressions to detect if a package has unit/integration tests or
performance tests (**).

(*): The NYC config now also uses the updated `testSuffix` value.
(**): As this file uses regular expressions, and the folders are defined
as globs, a conversion function from glob to regexp has been added. This
is ideally only temporarily and the `get-filters.js` script moves
towards using globs as well.