test(language-utils): no `groupName` in `NestedGroupMangleExpression`
Update the `NestedGroupMangleExpression` in preparation for deprecating
the `groupName` in favor of a provided capturing group string. This is
not yet fully implemented, but the expected behaviour is documented and
tested.

Two test cases are added to verify that the constructor of
`NestedGroupMangleExpression` throws when provided with a
`(sub)patternTemplate` without the provided capturing group string (and
no `groupName`, at least for the transition period). The scenarios are
updated in accordance with the new design. For each scenario, two test
cases are ran:

1. With the new constructor, omitting the `groupName` and using the
   provided capturing group string.
2. With the old constructor where the provided capturing group string is
   replaced by the string "(?<g>%s)", which matching the old scenarios.

The previously called "missing group" test scenario is replaced by a
scenario more accurately reflecting what it tests (also, "missing group"
is a nonsensical test scenario without a provided `groupName`). What it
actually tests is that when a group that is not the find and replace
capturing is triggered, the `NestedGroupMangleExpression` ignores that
match.

As the `groupName` is no longer relevant to the tests, the `group` field
of the `TestCase` types has been removed.

**NOTE**:
- The `.replace(/\s/g, "")` is present in most scenarios because the
  scenarios cannot assume that whitespace is removed from the template
  by the `NestedGroupMangleExpression`.