refactor(mangler-html-attributes): clean up the dependencies interface
Clean up the dependencies interface of the class. Now, all methods can
be called simply by providing the class options - lifting TypeScript to
ensure the dependencies only access what they need. This simplifies the
implementation, and making changes to the implementation in the future.

The implementations of the depenencies and their respective tests are
updated in accordance with the changes to the dependencies interface.

Fix incorrect indentation in all changed files following the formatting
rules.