test(language-html): improve embeds test suite based on mutants
Improve the unit test suites for the language-html package's embed
finders. For each of these this boils down to adding a test case for a
particular edge case that the Regular Expression is meant to catch.

For the style attribute embeds finder:
1. More than one space between the HTML tag name and the first attribute
   corresponding to a mutant at 8:66
2. An unquoted style attribute preceded by a valueless attribute
   corresponding to a mutant at 13:111
3. An unqouted style attribute in an element where the HTML tag is
   preceded by at least one space corresponding to a mutant at 13:57

For the style tag embeds finder:

First, the regular expression is improved by more accurately matching
(and not matching) attributes with values, especially those containing
">" in the closing style tag. All the added test relate to this change.

1. The attribute value in the closing tag contains a ">" in a double
   quoted attribute.
2. The attribute value in the closing tag contains a ">" in a single
   quoted attribute.
3. The closing tag isn't closed (missing ">"), three separate scenarios

For the script tag embeds finder:

The changes are identical to those for style tag embeds finder.