config: remove `process` as global
Remove the `process` object as an allowed global variable. Instead the
preferred way of using the `process` object is now by importing it
explicitly. This is in line with the Node.JS documentation [1].

While this only affects a relatively small number of files, the fix is
also trivial (just add an import statement/require for the `"process"`
module) so any violations *won't* be addressed here and should instead
be addressed over time as files containing violations change naturally.

--
1. https://nodejs.org/api/process.html#process