chore(cli): initialize reporters
Create a new directory that will be dedicated to reporters. Reporters
will be an abstraction that can be used to report on a WebMangler CLI
run. Specifically the statistics about the run. In essence, the
`logStats` function in `src/stats/log.ts` is a reporter (just not
conforming to the interface added here).