feat(types)!: require plugins to specify a character set
Update the definition of `MangleEngineOptions` (and by extension
`MangleOptions`) to make the `charSet` key non-optional - in-code
documentation is updated accordingly. This in turn requires that plugins
specify a character set for mangling explicitly (which all plugins in
this repository already do). This change is motivated by a recognition
that it's very hard to find a sensible & suitable character set that
works for any plugin. Hence, this change requires that each plugins
chooses its own character set.