chore(core)!: require plugins to specify a character set
In accordance with the `MangleEngineOptions`, the options parser of the
core's mangle engine now expects that plugins provide a character set
explicitly.

The test suites (unit and integration) for the core's engine are updated
accordingly. As now two test suites use a custom character set, commons
for the engine's tests have been added to expose some sample character
sets.