feat(scripts): run only unit tests in pre-commit hook
Update the pre-commit hook to only run unit tests to speed up the commit
hook. Unit tests are meant to be executed fast and validate behaviour at
a basic level. This, especially the former, makes it ideal for a
pre-commit hook, that should *quickly* validate the commit before it
happens.