refactor(cli): compute overall changed stats in stats module
Refactor the stats module of the CLI package to compute the overall
changed stats. For brevety the `FileStats` type is re-used, as the same
information is required at the aggregate level.

The implementation is taken from the existing computation of the overall
changed stats in `default-reporter.ts`. The existing test suite for the
stats computation is updated to test the aggregate values.

The test suites for other CLI modules that re-use the `ManglerStats`
type have been updated to *provide* the aggregate values, but they're
not tested yet because the corresponding implementations have not been
updated.

Miscellaneous changes:
- The size values in the `compute.test.ts` suite have been simplified to
  avoid having to test with numbers that have many digits after the
  decimal point.
- The `ManglerStats` type is changed to an interface because interfaces
  are now preferred over type defintions.