refactor(cli): use aggregate stats in the DefaultReporter
Update the `DefaultReporter` to print the provided aggregate stats
rather that the locally computed equivalent values, decoupling it from
the `getChangedPercentage` function of the stats module.

The test suite for the `DefaultReporter` is updated accordingly. In test
cases where the file *sizes* of the file entries are non-sensical (e.g.
because they're meant solely to test how percentages are printed), the
`aggregate` object is set to an object with equally non-sensical values.
This was done to avoid making unnecessary changes (w.r.t. what is
tested).