ci: compile packages before linting
Update the linting workflow to compile all packages prior to linting.
This is necessary because the `no-unresolved` rule of the
eslint-plugin-import [1] plugin will report unresolved imports for
internal dependencies if not all (relevant) packages are compiled. To
ensure all relevant packages are compiled when linting, all packages
will always be compiled prior to linting. The compiled packages are
cached to improve the runtime of the lint check.

Unfortunately, this creates an indirect build dependency on the
`compile` job of the "Validate Source Code" in that both jobs will fail
if compilation fails. This is something that may be addressed later and
is considered out of scope for this commit/branch.

--
1. https://github.com/import-js/eslint-plugin-import