refactor(cli): clean up stats types
Update the types.ts module of the stats directory to create a proper
type for `stats.aggregate` (nothing changed functionally due to the use
of a type alias).

Update the types.ts module of the reporters directory to use these new
types and avoid (potentially) confusing type aliases in the default
reporter's module. Also avoid the aliasing of `ManglerStats` in the
reporters directory's context to avoid potential confusion now - it is
always possible to redefine `ManglerStats` for reporters later.

Additionally, use a `ReadonlyMap` for `stats.files` to avoid reporters
changing the map.