ci: set up Stryker Dashboard
Configure Stryker and update the CI job running mutation tests to start
using Stryker Dashboard [1]. A cloud based view of Stryker mutation
test results.

The integration is configured such that each package will have it's own
dashboard separate from other packages. This is achieved using the
`dashboard.module` option [2]. By setting it to the first item in
`packagesList`, which is always the name of the package being mutation
tested in the CI, the correct behaviour is obtained.

--
1. https://dashboard.stryker-mutator.io
2. https://stryker-mutator.io/docs/stryker-js/configuration