config: fix jsdoc/match-description regular expressions
Fix the regular expression for the ESLint plugin eslint-plugin-jsdoc [1]
in accordance with a breaking change of v38.0.0 [2]. The description
regular expression now allows for a trailing newline (required since we
want a newline between the description and the tags, for readability).
A separate regular expression for param (`at-param`) descriptions is
introduced because there trailing newlines are not allowed.

--
1. https://github.com/gajus/eslint-plugin-jsdoc
2. https://github.com/gajus/eslint-plugin-jsdoc/releases/tag/v38.0.0