chore(cli): create CLI options type with reporters property
Create a CLI options type extending `WebManglerOptions` that adds
options unique to the CLI (compared to the Core). In this instance it
adds the `reporters` options. This option is optional with a default
value available later.

The optional option requires the addition of a separate type defining
the actual configuration - without any optional values - derived from
the options provided to the CLI.

Additionally, the type of the reporters option is taken from the already
existing `Reporter` type of the CLI. This type did need to be exported
first.