refactor(cli): Use CliOptions and CliConfig types for loading configs
Update the configuration loading logic to load configuration files as
`WebManglerCliOptions`, and return configs as `WebManglerCliConfig`.

The corresponding test suites have been updated accordingly. Note that
the added test for the default configuration is skipped because, as of
this commit, the default configuration *cannot* provide a reporter (due
to the `DefaultReporter` having the wrong constructor signature).