refactor(cli): Update `Reporter::report` to require a Writer
Update the method signature of the `.report` method of the `Reporter`
interface to require a `Writer` to be provided. This is motivated by the
fact that the `DefaultReporter` (or any reporter for that matter) must
be instantiable without providing a writer, because it must be
instantiable in the _WebMangler_ configuration file.

The `DefaultReporter` is updated accordingly, including the constructor.
The `New` method has not been updated (considered out of scope), but a
lint error in the implementation is suppressed with a comment - the
function should be updated/removed in a followup commit.

The existing usage of the `Reporter` interface in the project (source &
tests) has been updated accordingly.