config: enable imports/group-exports rule
Enable the imports/group-exports rule [1] to force grouping exports, and
as a result force consistent exporting style - grouping all exports
together at the end of a file.

In 224a48365166e31ae1dd0f6c8c1836d644810c43 the rule was added and
explicitly disabled because

> [The rule] is disabled because I currently feel that both grouping and
> not grouping has merit. If this turns out to be pain point in the
> futurethe rule can be enabled.

Now, the realization came that grouping exports in this way forces a
consistent exporting style. Hence, enforcing it prevents back-and-forth
in one file, as well as discussion on the topic during coding or review.

One file currently using ungrouped exports was modified (partly) to
demonstrate the effectiveness of the rule. Running the linter on this
commit will result in a linter warning from the imports/group-exports
rule in the edited file.

--
1. https://github.com/import-js/eslint-plugin-import/blob/HEAD/docs/rules/group-exports.md