refactor(language-css): initial work for embeds
Make initial changes to the language-css package to be able to support
using embeds. The added functionality is tested (where tests do not yet
apply, they're currently skipped), but does not do anything useful yet.

The current plan for embeds in the language-css package is to extract,
e.g., media queries as embeds. This would allow for more precise regular
expressions by adding the requirement that, e.g., CSS properties appear
only within a curly-brace pair (this would actually be re-adding, but it
was removed to support mangling in media queries [1]).

--
1. https://github.com/ericcornelissen/webmangler/issues/127