refactor(core,types): change `Iterable` to `Collection`
Replace the use of the type `Iterable` from the WebMangler core API by
a newly defined type: `Collection`. The intend of the `Collection` type
is to provide support for types that are collections of other types.
This is similar to the `Iterable` type, but excludes instances of
`Iterable` that can only be iterated once.

This changeset was chosen as it's the minimal number of changes required
to have the WebMangler core API use `Collection` instead of `Iterable`.

Additionally, `embeds.ts` has its exports moved to the end of the file
in accordance with linter rules (because it is being changed anyway).