chore: create initial GitHub issue template for questions
Add an initial GitHub issue form template for questions. This template
is intended to be used when the issue creator simply doesn't know
something, e.g. how to achieve a certain outcome. The template is very
simple, basically just asking the submitter to write their question and
also, if possible, specify the packages they're asking the question
about.