fix(scripts): running scripts on Windows
On Windows, commands passed to the custom `execSync` function in scripts
is not detected as is without setting `shell: true`. This is based on
the child_process documentation, specifically the section "Spawning .bat
and .cmd files on Windows" on [1].

`shell` is set to override the provided `options` as it is not expected
that callers ever need to specify the shell. We also want to avoid
accidentally overriding `shell` with a falsy value when calling
`execSync` (without realizing the negative consequences this has on
Windows development).

--
1. https://nodejs.org/api/child_process.html