refactor(language-utils): object-as-parameter for simple language plugin
Update the `SimpleLanguagePlugin` constructor to accept an object as
parameter and deprecate the existing constructor. The test suite for the
class now runs the same tests for both version of the constructor. The
documentation is already updated for use with the new constructor to
encourage users to upgrade.

Additionally, the test for `SimpleLanguagePlugin` are updated in line
with linter warnings/errors.