test(cli): initial suite for the JsonReporter class
Create an initial suite for the (new) `JsonReporter` class for reporting
mangling stats in JSON format. The suite tests some basic expected
functionality of the reporter, namely the expected key-value pairs in
various scenarios.

Note: the existing `JsonReporter` is not yet passing this suite due to a
mistake in the serialization of the `files` key of the `ManglerStats`.