test(core): improve test suite for embeds.ts
Improve the test suite for embeds.ts to kill some mutants found by
Stryker.

First is an improvements to the tests for `getEmbeds` to test for the
exact contents of the returned `file.content`. This catches potential
problems with the string builder, such as it being initialized to a
non-empty array (= the mutant that identified the problem).

Second is a specific test case for `reEmbed` that checks the early exit
happens for the "no embeds" case. This case is needed specifically when
there's instance of the id prefix present in the file contents because
this will cause a `TypeError` on line 154 of `embeds.ts`.