refactor(benchmarking): prefer readonly types
Update the benchmarking package to use both `Readonly` variants of
collection types (e.g. `ReadonlyArray`) and `readonly` properties. In
the source code this ensures data isn't unintionally changed by the
function that the collection (sometimes property) is provided to. This
treatment is also applied to tests as we want to avoid accidentally
changing values from `TestScenarios`/`TestCase`s.

Notably, this required an implementation change to `math.ts::medianOf`
because, by using `[].sort`, it altered the order of the provided array.

A notable field that wasn't made `readonly` is the `medianDuration` of
`BenchmarkStats` (`stats/types.ts`). This is because 1) the interface is
exported, so making it `readonly` would be a breaking change; and 2) it
does not make sense to restrict the receiver from changing it, as there
is no harm in it. The type is not used as an input to the package. (This
argument also applies to the other case where a type was intentionally
*not* made readonly, in `run.ts` at line 60.)