test(core): improve test suite for engine/replace.ts
Improve the test suite for the replacement module of the engine by
rewriting the stubs such that they return the expected value based on
the actual provided value. Doing this ensures the expected values are
passed around inside the function correctly, as opposed to returning
the expected value regardless of whether or not the inputs make sense.

Also fix the assertion within the test. Prior to this change the check
would always pass, regardless of the value of the contents of any of the
files. This is because the replacement function works in place, hence
the values in the expected array would be identical in some cases where
the actual and expected file object were the same in the test definition
(e.g. because the output file is expected to be identical to the input
file).