ci(scripts): update mutation testing filter
Update the logic creating the filter for mutation testing from checking
the list of packages that should be mutation tested to building a list
of packages that have **unit** tests. It checks for unit tests because
mutation testing is only performed on the unit tests (see
`stryker.config.js:L25`).