feat(cli): add support for petty-printing to `JsonReporter`
Update the `JsonReporter` to support pretty-printing the JSON report. To
achieve this, the `JsonReporter` now has a constructor that takes an
(optional, at least for backwards compatability) parameters object. This
object may have a (boolean) `prettyPrint` entry which, if `true`, causes
the reporter to format the JSON for humans.

The chosen formatting is based on the MDN documentation for
`JSON.stringify` [1] which states:

    Using a tab character mimics standard pretty-print appearance[.]

Contrary to the MDN documentation, the `space` argument (used for
formatting) is set to `undefined` (instead of `null`) if the JSON should
not be pretty-printed. The reason for this "choice" is that TypeScript
does not recognize `null` as a valid value for this argument. Based on
the fact that `space` is an optional argument, and from local testing,
setting it to `undefined` is fine as well.

--
1. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify