test(language-html): don't detect embeds consisting of only whitespace
Update the test suite for all embeds finders of the language-html
package to omit the test that verifies embeds consisting of only
whitescapce are returned and replace it by tests that verify the
opposite (no embed is returned). For completeness, tests are created for
two different whitespace characters to reduce the chances of the
implementation checking for a specific one.