feat(language-html): don't detect embeds consisting of only whitespace
Update the logic for decoding RegExp matches of embeds for the package
to look for matches consisting of only whitespace (`^\s*$`) and don't
return those matches as an embed.

It was considered to update the `REGEXP_SOMETHING` expression to do this
filtering but all attempts resulted in bugs w.r.t. the end of the embed
being skipped because the embed-terminating character would match the
non-whitespace character of the expression.