refactor(language-css): use dependency injection for CssClassMangler
Refactor the `CssClassMangler` class to use (explicit) dependency
injection for the helper methods. The approach taken is based on that of
other mangler plugins. The documentation added in `index.ts` comes
(almost) straight from `class.ts`.

The test suites for the class have been updated accordingly. The unit
test suite is based on the old integration test suite as well as the
unit test suite for other mangler classes. The integration test suite is
based entirely of that from other mangler classes.