refactor(mangler-css-classes): make options fields `readonly`
Update the `CssClassManglerOptions` so that all options are treated as
`readonly` bye the `CssClassMangler`. This does not affect users of this
package, it just provides them a guarentee that the object won't get new
value references.

The version update of `CssClassManglerOptions` is updated accordingly.
Note the TSDoc changes for `ignoreClassNamePattern` include the addition
of the since tag (based on the CHANGELOG.md for the core package, line
31) that was missing before.