refactor!(language-utils): use readonly types in `MultiLanguagePlugin`
Update the `MultiLanguagePlugin` to use readonly-variant of types as
much as possible (and makes sense). This commit is focused to the parts
of the class that are private or otherwise independent of the
`WebManglerLanguagePlugin` type.

Notably, the `::getEmbeds`' argument is now wrapped in the generic
`Readonly<T>` type. This ensures the file contents aren't changed by the
implementation. This change may be carried over to the types package in
the future (though that would be a breaking change there).

Due to the breaking change (see below) the test suite for the class had
to be slightly altered.

BREAKING CHANGE: The `MultiLanguagePlugin` no longer explicitly works
with all iterable types. Notably, the intention is to drop support for
generators only.