refactor!(language-utils): use readonly types in `SimpleLanguagePlugin`
Update the `MultiLanguagePlugin` to use readonly-variant of types as
much as possible (and makes sense). This commit's focus is primarily
on the parts of the class that are private or otherwise independent of
the `WebManglerLanguagePlugin` type.

BREAKING CHANGE: The `EmbedsGetter` has been updated to require that the
implementation treats the `WebManglerFile` it operators on as readonly.
In particular, the `contents` property can no longer be edited by an
`EmbedsGetter` implememtation.