refactor(cli): move `RawStatsData` to to types.ts
Move the `RawStatsData` type of the stats module of the cli package to
types.ts to make it re-usable. This is in preparation for making it more
re-usable and make composition of compute.ts and paths.ts easier.

In conjuction, the interface is updated to make the `inFiles` and
`outFiles` array contents readonly (not just the object property). This
should help ensure the `computeStats` has no unintended side effects on
the input data.