feat(cli): only use relative file paths in reports
Update the (exported) `computeStats` function to omit the full path from
the stats about an individual file in favor of a relative path from the
project root. The intent is to avoid unintentional leakage of sensitive
information that might be present in the full path. As an added bonus it
(may) simplify the output of reports to be more human-readable.

The implementation currently wraps the pre-existing `computeStats`
function so as not to touch it (and unintionally make a wrong change).
Unfortunately the original implementation isn't very extensible so this
approach was chosen instead.