refactor!(mangler-utils): use readonly types in `MultiManglerPlugin`
Update the `MultiManglerPlugin` to use readonly-variant of types as much
as possible (and makes sense). This commit is focused to the parts of
the class that are private or otherwise independent of the
`WebManglerPlugin` type.

Due to the breaking change (see below) the test suites for the class had
to be slightly altered.

BREAKING CHANGE: The `MultiManglerPlugin` no longer explicitly supports
all `Iterable` types.