refactor(language-utils): use readonly types in `MangleExpression` impl
Update both implementations of the `MangleExpression` interface to use
the readonly array variant when working with Regular Expression matches
to prevent unexpected side-effects. This has no effect to the outside
world. Also, the `RegExpMatchGroups` type is now shared between the two
implementations.

All changes to the in-code documentation are forced by the ESLint rules.
Note that the at-since tag was removed where at-inheritDoc is present
because it's carried over from the inherited docs.