feat(types): add `caseSensitive` flag to certain `MangleExpression` opts
Update the `CssDeclarationValueOptions` and `QuerySelectorOptions`
mangle expression options with an optional `caseSensitive` option
to control the case sensitivity of CSS decleration values and query
selectors (resp.) in line with the case-sensitivity bug for HTML
attribute mangling. The changed options types are based on the
`MangleExpressionOptions` found in the `language-options.ts` module of
the mangler-html-attributes package.