test(core): detected nested embeds
Create a new test scenario (called "nested embeds") for the embeds
module of the core package to test if it can detect nested embeds. This
scenario currently consists of one case with a stylesheet embed
containing another embed (a media query).

To be able to properly test this new scenario the test assertions have
to be updated. In particular, the embeds content (and by extension raw
value) must be checked with a `.to.match` as its content could now
contain a embed id (similar to the `file.content`).

As a result of this, as well as the realization that always returning
values from the `WebManglerLanguagePluginMock.getEmbeds` method is not
ideal, the test scenario called "sample" is updated. The test suite
should be functionally ~equivalent to what it was prior to this commit.