feat(core): support nested embeds
Add support for detected embeds nested in other embeds. The
implementation is recursive to increase re-used and satisfies[^1] the
pre-existing tests for nested embeds.

To support the recursive calls, the use of `Iterable` for
`IdentifiableWebManglerEmbed` had to be replaced by `Collection` in line
with the type of the `files` for the `getEmbeds` function.

[^1]: The change to `embeds.test.ts` is required as `calledOnceWith`
verifies the stub was called only once in total, not once for the
expected argument(s).