chore(core): update performance tests for embeds
Update the performance tests for embeds to account for the now-recursive
nature of the `getEmbeds` function. With the existing stub the test
would end up in an endless recursion spiral. To fix this the language
plugin mock is no configured to return embeds only once.

The budget for this performance has been increased due to the
now-recursive nature of the function. Especially with the configured 100
top-level embeds (see line 17 of the suite), this adds up w.r.t.
performance. Hence, the budget is doubled.