chore: upgrade the npm `lockfileVersion` from 2 to 3
Update the version of the lockfile (`package-lock.json`) from 2 to 3
(latest at the time of commiting). This is done to be more up-to-date
as well as reduce the size of the lockfile. To accomidate this, the
minimum npm version (defined in the root manifest, `./package.json`)
was increased from >=7.0.0 to >=8.1.2 in line with the following line
from the 8.1.2 release notes:

> allow `--lockfile-version` config to be string and coerce to number

Which is required for the last change found in this commit: the addition
of the "lockfile-version" option to the .npmrc configuration file. This
instructs npm to use lockfile version for `package-lock.json` that is
found in the configuration file - forcing contributors to use this
version when running `npm install` for this project.