test(core): re-embedding nested embeds
Add an integration test suite for embeds that tests whether using
`getEmbeds` and `reEmbed` in sequence as intended results in the files
being changed correctly. This is a useful test in general, and is
classified as 'integration' because it checks whether to two functions
integrate correctly. However, it is also created to test against a gap
in the implementation, namely that nested embeds aren't re-embedded
properly. Hence, this commit has one failing test case.