fix(core): re-embedding nested embeds
Update the `getEmbeds` logic to add nested embeds before "top-level"
embeds to ensure they're re-embedded prior to their "top-level" counter
parts. If not, this results in the nested embed's re-embedding not
appearing in the parent's re-embedding.

Also, adjust the unit tests for the embeds logic to accomodate for the
change in the implementation.