ci: always upload test coverage to Codecov
Update the CI configuration to upload test coverage results to Codecov
even if (A) tests failed or (B) test coverage is insufficient (in which
case the `Run test suite for packages/xxx` would fail and error). This
should give a more complete overview of test coverage over time as more
data will be available on Codecov. This is particularly interesting in
the case of scenario B.