chore: set up support for an end-to-end test category
Update the test script and Mocha configuration to support an end-to-end
(e2e) test type. The definition (of sorts) of e2e tests is that they
span packages in the sense that they require a package is not referenced
in the package's TypeScript configuration - hence requiring all packages
to be compiled prior to starting the tests.

The process import added to the Mocha configuration is in line with
ESLint errors for that file.