ci: run end-to-end tests
Update the `code-checks.yml` workflow file to run end-to-end tests if
and only if the code compiles and unit+integration tests pass. The
packages for which to run e2e tests is determined by the newly created
filter which scans the package for any `.test.ts` file inside a
`__tests__/e2e` folder.

For now, e2e tests will be run in the CI without coverage, as coverage
is not yet sufficient.