test(language-css): don't match selectors in rulesets
Update the test suite for the `MangleExpression`s of CSS selectors to
require that they no longer match query selectors inside of rulesets.
This change follows from the recent change of handling media queries
as embeds (item "Treat media queries as embeds." from the changelog).

For `query-selector.test.ts`, one test scenario was removed as it's
effectively identical to the test scenario altered in this commit.